Add 1mm backoff when determining z down limit switch margin to PCB surface
Regarding the nozzle height calibration and "probe down functions," the default "0mm" backoff was not enough to actually move away from the limit switch and stop pressing it. Without this change, we were getting a limit switch error during nozzle height calibration when we tried to move the nozzle up to find the "difference to just touching." This does require an extra 1mm compression in the tube but we just couldn't get it to work otherwise.